validate cart, update view cart

+++ src/main/java/service/impl/GioHangService.java
@@ -6,6 +6,7 @@
 package service.impl;
 
 import entity.GioHangEntity;
+import entity.SanPhamEntity;
 import java.util.ArrayList;
 import java.util.HashMap;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,7 +28,7 @@ public class GioHangService implements IGioHangService {
 
     @Autowired
     IKhachHangRepository khRepository;
-    
+
     @Autowired
     ISanPhamRepository spRepository;
 
@@ -36,42 +37,70 @@ public class GioHangService implements IGioHangService {
         if (khRepository.findOne(cart.getMakh()) == null) {
             return "Không tìm thấy khách hàng";
         } else {
-            int magiohang = ghRepository.isExists(cart.getMasp());
-            if (magiohang < 0) {
-                int giatong = cart.getGia() * cart.getSoluong();
-                cart.setGiatong(giatong);
-                if (ghRepository.create(cart) == true) {
-                    return "Thêm giỏ hàng thành công";
-                } else {
-                    return "Thêm giỏ hàng thất bại";
+            int magiohang = ghRepository.isExists(cart.getMasp(), cart.getMakh());
+            SanPhamEntity sp = spRepository.findById(cart.getMasp());
+            int soluong = sp.getSoLuong() - cart.getSoluong();
+            if (spRepository.updateSoluong(sp.getMaSP(), soluong) == true) {
+                if (magiohang < 0) {
+                    int giatong = cart.getGia() * cart.getSoluong();
+                    cart.setGiatong(giatong);
+                    if (ghRepository.create(cart) == true) {
+                        return "Thêm giỏ hàng thành công";
+                    } else {
+                        return "Thêm giỏ hàng thất bại";
+                    }
                 }
-            } else {
                 if (ghRepository.incremental(magiohang, cart.getGia(), cart.getSoluong()) == true) {
                     return "Thêm giỏ hàng thành công";
                 } else {
                     return "Thêm giỏ hàng thất bại";
                 }
+            } else {
+                return "Thêm giỏ hàng thất bại";
             }
         }
     }
+
     @Override
-    public String updateCart(GioHangEntity cart){
-        if(spRepository.checkQuantity(cart.getMasp(), cart.getSoluong()) == true){
-            if (ghRepository.update(cart) == true) {
-                return "Cập nhật số lượng thành công";
+    public String updateCart(GioHangEntity cart) {
+        SanPhamEntity sp = spRepository.findById(cart.getMasp());
+        GioHangEntity gh = ghRepository.findById(cart.getMagiohang());
+        int soluong = sp.getSoLuong();
+        if (cart.getSoluong() - gh.getSoluong() == 1) {
+            soluong = soluong - 1;
+        } else if(cart.getSoluong() - gh.getSoluong() == -1){
+            soluong = soluong + 1;
+        }else if(cart.getSoluong() - gh.getSoluong() < -1){
+            soluong = soluong + (gh.getSoluong() - cart.getSoluong()); 
+        }else if(cart.getSoluong() - gh.getSoluong() > 1){
+            soluong = soluong - (cart.getSoluong() - gh.getSoluong()); 
+        }
+        if (soluong < 1) {
+            return "Số lượng đã vượt quá số lượng tồn";
+        } else {
+            if (spRepository.updateSoluong(sp.getMaSP(), soluong) == true) {
+                if (ghRepository.update(cart) == true) {
+                    return "Cập nhật số lượng thành công";
+                } else {
+                    return "Cập nhật số lượng thất bại";
+                }
             } else {
                 return "Cập nhật số lượng thất bại";
             }
-        }
-        else{
-            return "Số lượng vượt quá số lượng tồn";
         }
     }
 
     @Override
     public String deleteCart(int magiohang) {
-        if (ghRepository.delete(magiohang) == true) {
-            return "Xóa giỏ hàng thành công";
+        GioHangEntity gh = ghRepository.findById(magiohang);
+        SanPhamEntity sp = spRepository.findById(gh.getMasp());
+        int soluong = sp.getSoLuong() + gh.getSoluong();
+        if (spRepository.updateSoluong(sp.getMaSP(), soluong) == true) {
+            if (ghRepository.delete(magiohang) == true) {
+                return "Xóa giỏ hàng thành công";
+            } else {
+                return "Xóa giỏ hàng thất bại";
+            }
         } else {
             return "Xóa giỏ hàng thất bại";
         }
@@ -81,9 +110,9 @@ public class GioHangService implements IGioHangService {
     public ArrayList<GioHangEntity> getAll(int makh) {
         return ghRepository.findAll(makh);
     }
-    
+
     @Override
-    public int totalPrice(int makh){
+    public int totalPrice(int makh) {
 //        int total = 0;
 //        int total = ghRepository.totalPrice(makh);
 //        for(GioHangEntity cart : cartList){
